refactor(account-menu): contain notifications block; drop redundant desktop ProfileButton
- NotificationList renders as a self-contained bounded card (own border/bg,
  scrolls within max-h) instead of an unseparated inline block, so it reads as
  one section clearly divided from the account nav items.
- SidebarAccount: notifications now sit between the nav targets and the
  theme/logout row.
- Desktop titlebar no longer renders its own ProfileButton account menu; the
  shared sidebar SidebarAccount is the single account menu on desktop too.
  Relocate NAV_ITEMS into Header/menuData.tsx and delete the now-unused
  ProfileButton.tsx and orphaned ThemeToggleButton.tsx.
